+++ client/node/BUILD.bazel
@@ -10,7 +10,6 @@ go_library(
         "//client/mainchain:go_default_library",
         "//client/notary:go_default_library",
         "//client/observer:go_default_library",
-        "//client/p2p:go_default_library",
         "//client/params:go_default_library",
         "//client/proposer:go_default_library",
         "//client/simulator:go_default_library",
@@ -20,6 +19,7 @@ go_library(
         "//shared:go_default_library",
         "//shared/cmd:go_default_library",
         "//shared/debug:go_default_library",
+        "//shared/p2p:go_default_library",
         "@com_github_ethereum_go_ethereum//node:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@com_github_urfave_cli//:go_default_library",
