+++ client/notary/BUILD.bazel
@@ -12,8 +12,8 @@ go_library(
         "//client/contracts:go_default_library",
         "//client/database:go_default_library",
         "//client/mainchain:go_default_library",
-        "//client/p2p:go_default_library",
         "//client/params:go_default_library",
+        "//shared/p2p:go_default_library",
         "@com_github_ethereum_go_ethereum//accounts:go_default_library",
         "@com_github_ethereum_go_ethereum//accounts/abi/bind:go_default_library",
         "@com_github_ethereum_go_ethereum//core/types:go_default_library",
