+++ client/observer/BUILD.bazel
@@ -8,9 +8,9 @@ go_library(
     deps = [
         "//client/database:go_default_library",
         "//client/mainchain:go_default_library",
-        "//client/p2p:go_default_library",
         "//client/syncer:go_default_library",
         "//client/types:go_default_library",
+        "//shared/p2p:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
     ],
 )
@@ -22,10 +22,10 @@ go_test(
     deps = [
         "//client/database:go_default_library",
         "//client/mainchain:go_default_library",
-        "//client/p2p:go_default_library",
         "//client/params:go_default_library",
         "//client/syncer:go_default_library",
         "//client/types:go_default_library",
+        "//shared/p2p:go_default_library",
         "@com_github_sirupsen_logrus//hooks/test:go_default_library",
     ],
 )
