+++ client/proposer/BUILD.bazel
@@ -11,13 +11,13 @@ go_library(
     deps = [
         "//client/database:go_default_library",
         "//client/mainchain:go_default_library",
-        "//client/p2p:go_default_library",
         "//client/params:go_default_library",
         "//client/syncer:go_default_library",
         "//client/txpool:go_default_library",
         "//client/types:go_default_library",
         "//proto/sharding/v1:go_default_library",
         "//shared/legacyutil:go_default_library",
+        "//shared/p2p:go_default_library",
         "@com_github_ethereum_go_ethereum//accounts:go_default_library",
         "@com_github_ethereum_go_ethereum//accounts/abi/bind:go_default_library",
         "@com_github_ethereum_go_ethereum//core/types:go_default_library",
