+++ client/simulator/BUILD.bazel
@@ -7,10 +7,10 @@ go_library(
     visibility = ["//client:__subpackages__"],
     deps = [
         "//client/mainchain:go_default_library",
-        "//client/p2p:go_default_library",
         "//client/params:go_default_library",
         "//client/syncer:go_default_library",
         "//proto/sharding/v1:go_default_library",
+        "//shared/p2p:go_default_library",
         "@com_github_ethereum_go_ethereum//event:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
     ],
@@ -23,9 +23,9 @@ go_test(
     embed = [":go_default_library"],
     deps = [
         "//client/mainchain:go_default_library",
-        "//client/p2p:go_default_library",
         "//client/params:go_default_library",
         "//client/types:go_default_library",
+        "//shared/p2p:go_default_library",
         "@com_github_ethereum_go_ethereum//:go_default_library",
         "@com_github_ethereum_go_ethereum//accounts/abi/bind:go_default_library",
         "@com_github_ethereum_go_ethereum//common:go_default_library",
