+++ client/syncer/BUILD.bazel
@@ -11,10 +11,10 @@ go_library(
     deps = [
         "//client/database:go_default_library",
         "//client/mainchain:go_default_library",
-        "//client/p2p:go_default_library",
         "//client/params:go_default_library",
         "//client/types:go_default_library",
         "//proto/sharding/v1:go_default_library",
+        "//shared/p2p:go_default_library",
         "@com_github_ethereum_go_ethereum//accounts/abi/bind:go_default_library",
         "@com_github_ethereum_go_ethereum//common:go_default_library",
         "@com_github_ethereum_go_ethereum//event:go_default_library",
@@ -34,10 +34,10 @@ go_test(
         "//client/contracts:go_default_library",
         "//client/database:go_default_library",
         "//client/mainchain:go_default_library",
-        "//client/p2p:go_default_library",
         "//client/params:go_default_library",
         "//client/types:go_default_library",
         "//proto/sharding/v1:go_default_library",
+        "//shared/p2p:go_default_library",
         "@com_github_ethereum_go_ethereum//accounts:go_default_library",
         "@com_github_ethereum_go_ethereum//accounts/abi/bind:go_default_library",
         "@com_github_ethereum_go_ethereum//accounts/abi/bind/backends:go_default_library",
