+++ client/txpool/BUILD.bazel
@@ -6,7 +6,7 @@ go_library(
     importpath = "github.com/prysmaticlabs/prysm/client/txpool",
     visibility = ["//client:__subpackages__"],
     deps = [
-        "//client/p2p:go_default_library",
+        "//shared/p2p:go_default_library",
         "@com_github_ethereum_go_ethereum//event:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
     ],
